cloud_storage/tests: refactor topic_recovery_test utils to own module
